requirements.txt
torch 
torchvision 
torchaudio 
fair-esm 
requests 
scikit-learn 
matplotlib 
tqdm
Bio